functions/delete_campaign/requirements.txt
functions-framework
google-cloud-firestore
google-cloud-storage
firebase-admin
